12-capstones/smartcity/03-analytics/producer-cdk/requirements.txt
-e .
pytest
aws-cdk.aws-kinesis
aws-cdk.aws_kinesisfirehose